Port to contiki-ecc
Contiki-ecc is a port of TinyECC to contiki (thus porting the nesC code
to C). At this stage, the contiki-ecc code has been heavily modified to
support the ECC certificates. Also, it is unclear if the contiki-ecc
code that I downloaded is in a working condition (some trivial error
would prevent the code from compiling).

The tools have been proved to compile, but have not been tested for
correctness.